requirements.txt
streamlit
pylast
bcrypt
mysql-connector-python
python-dotenv
requests

